move the picker to jackdaw_feathers, make the picker a styled widget
