+++ crates/jackdaw_feathers/src/lib.rs
@@ -14,6 +14,7 @@ pub mod list_view;
 pub mod menu_bar;
 pub mod panel_header;
 pub mod panel_section;
+pub mod picker;
 pub mod popover;
 pub mod progress;
 pub mod scroll;
@@ -61,6 +62,7 @@ impl Plugin for EditorFeathersPlugin {
             context_menu::plugin,
             panel_header::plugin,
             tooltip::plugin,
+            picker::plugin,
         ));
     }
 }
