+++ crates/jackdaw_fuzzy/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
 name = "jackdaw_fuzzy"
-version = "0.1.0"
+version = "0.4.0"
 edition = "2024"
+description = "Internal crate for the jackdaw editor"
+repository = "https://github.com/jbuehler23/jackdaw"
+license = "MIT OR Apache-2.0"
+include = ["Cargo.toml", "LICENSE*", "fonts/**/*", "src/**/*"]
 
 [dependencies]
-bevy.workspace = true
-bevy_ui_text_input.workspace = true
-jackdaw_feathers.workspace = true
 nucleo-matcher = "0.3.1"
 
 [lints]
