--- crates/jackdaw_fuzzy/src/prelude.rs
@@ -1,5 +1,0 @@
-pub use crate::matcher::{FuzzyItem, Match, MatchedStr};
-pub use crate::picker::{
-    FuzzyPicker, PickerEntities, PickerInputOf, PickerListOf, PickerPlugin,
-    RegisterFuzzyItemAppExt as _, SelectInput, SpawnItemInput, WithPickerInput, WithPickerList,
-};
